backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9

# Async DB driver (optional, for async SQLAlchemy)
asyncpg==0.29.0

# Cache
redis==5.0.8

# Task queue
celery==5.4.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & config
pydantic[email]==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# HTTP & file handling
python-multipart==0.0.9
httpx==0.27.2
pillow==10.4.0
aiofiles==24.1.0

# WebSocket
python-socketio==5.11.4

# AI / RAG
langchain==0.3.1
langchain-community==0.3.1
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
sentence-transformers==3.1.1
pypdf==4.3.1
ollama==0.3.3

# Utilities
python-dateutil==2.9.0
